Keep default values raw (no parsing)
We do not care what default value is provided in JSON schema.
Valid or invalid, matters not.

Two cases for future useage:
1. Use default value to encode missing fields in avro:in()
   when encoding avro JSON or binary stream.
2. Use default value to up-vert or down-vert decoded avro:out()

For case #1, the encoder will crash if invalid value is found
in schema, user will have to fix schema anyway.

For case #2, invalid value will be exposed to higher level, then
it's up to the caller to decide either fix code or schema.