For string types, accept atoms as values
This PR adds functionality that allows string fields to accept atoms as
values. I'm not sure if this is a controversial in Erlang, but in Elixir
it is very common to encode atom values as strings. For example,
https://hexdocs.pm/jason/Jason.html#encode/2 happily encodes atom values
as strings.

This has come up several times in projects where we are storing values
as atoms, and we have to rewrite the value before passing it to the Avro encoder.

- [  ] Properly encode union types (Test is broken)