Do not convert boolean atoms to strings
The boolean atoms 'true' and 'false' should be skipped when converting
to a string. Otherwise, when handling a union type which specifies
strings before booleans, the values 'true' or 'false' will get turned to
a string, which corrupts data at serialization time.

Since the null and nil atoms are already skipped to avoid type
confusion, it seems reasonable to extend the exception to booleans. Most
encoding of strings should ideally be done using binaries in accordance
to the documentation, which specifies 'iolist()' as the type for strings
and atoms appear to be undocumented.

This current issue wasn't found through any regular test in this
repository, but was found in property-test in a private corporate
project. Without sharing similar schemas it is a bit complex to extend
the existing tests for an undocumented feature, so I skipped it for now.